refactor: convert CLI commands from static classes to functional approach
- Convert AgentsCommands, TuiCommand, WorktreeCommands, and ComposerCli from static classes to plain functions
- Remove withCli helper and use Effect.gen directly in handlers
- Improve testability by eliminating static method complexity
- Rename files to use kebab-case (agents.ts, tui.ts, git-worktree.ts, composer.ts)
- Rename service files to use kebab-case (agent-cli.ts, git-worktree-cli.ts)
- Update all imports to use new functional exports
- Preserve service types and layer structure for ComposerCli
- Fix formatting issues in git-worktree-cli.ts
- Update test files to use correct import names
- Maintain backward compatibility by exporting both old and new names
