+++ packages/cli/src/lib/index.ts
@@ -1,4 +1,1 @@
-export * from "@open-composer/agent-router";
-export { AgentCli } from "../services/AgentCli.js";
-export { WorktreeCli } from "../services/WorktreeCli.js";
 export * from "./cli.js";
